Add in-memory LRU cache
