+++ go.mod
@@ -6,6 +6,7 @@ require (
 	github.com/PuerkitoBio/purell v1.1.1
 	github.com/PuerkitoBio/urlesc v0.0.0-20170810143723-de5bf2ad4578 // indirect
 	github.com/andybalholm/brotli v1.0.1
+	github.com/hashicorp/golang-lru v0.5.4
 	github.com/rs/zerolog v1.21.0
 	github.com/stretchr/testify v1.7.0
 	go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp v0.19.0
@@ -16,6 +17,6 @@ require (
 	go.opentelemetry.io/otel/trace v0.19.0
 	golang.org/x/net v0.0.0-20210331212208-0fccb6fa2b5c
 	golang.org/x/sync v0.0.0-20210220032951-036812b2e83c
-	golang.org/x/text v0.3.3
+	golang.org/x/text v0.3.6
 	google.golang.org/grpc v1.36.1
 )
