+++ httphandler/httphandler.go
@@ -13,7 +13,7 @@ import (
 )
 
 // New creates a new Handler.
-func New(resolver *urlresolver.Resolver) *Handler {
+func New(resolver urlresolver.Resolver) *Handler {
 	return &Handler{
 		resolver: resolver,
 	}
@@ -21,7 +21,7 @@ func New(resolver *urlresolver.Resolver) *Handler {
 
 // Handler is an HTTP request handler that can resolve URLs.
 type Handler struct {
-	resolver *urlresolver.Resolver
+	resolver urlresolver.Resolver
 }
 
 func (h *Handler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
